更新数据模型和定时任务调度器：将 escrow_amount 字段类型更改为 DECIMAL(12, 2) 以提高精度；优化过期积分处理逻辑，使用原子 SQL 操作避免并发问题，添加余额不足的处理；增强任务调度器，支持高优先级任务执行和健康检查机制，记录任务执行耗时和状态。

+++ backend/app/models.py
@@ -203,7 +203,7 @@ class Task(Base):
     is_paid = Column(Integer, default=0)  # 1=paid, 0=not paid
     payment_intent_id = Column(String(255), nullable=True)  # Stripe Payment Intent ID，用于关联支付记录
     payment_expires_at = Column(DateTime(timezone=True), nullable=True)  # 支付过期时间（待支付状态的任务）
-    escrow_amount = Column(Float, default=0.0)
+    escrow_amount = Column(DECIMAL(12, 2), default=0.00)  # 托管金额（任务金额 - 平台服务费）
     is_confirmed = Column(Integer, default=0)  # 1=confirmed, 0=not
     paid_to_user_id = Column(String(8), ForeignKey("users.id", ondelete="SET NULL"), nullable=True)  # 删除用户时设为NULL
     is_public = Column(Integer, default=1)  # 1=public, 0=private (仅自己可见)
